dawn.node: Implement the [Clamp] and [EnforceRange] WebIDL attributes.
These are implemented by wrapping the integer types in transparent
ClampedInteger<> and EnforceRangeInteger<> structures.

Some parts of the core needed to be updated after this, either to
disambiguate conversions, or because of bugs (u32 vs u64).

To make the CTS tests checking for this pass, the errors returned when
conversion FromJS failed needed to be updated to TypeError and not just
the generic Napi::Error.

Bug: dawn:1123

+++ src/dawn/node/binding/GPURenderPassEncoder.cpp
@@ -202,7 +202,7 @@ namespace wgpu::binding {
         Converter conv(env);
 
         wgpu::Buffer b{};
-        uint32_t o = 0;
+        uint64_t o = 0;
 
         if (!conv(b, indirectBuffer) ||  //
             !conv(o, indirectOffset)) {
@@ -218,7 +218,7 @@ namespace wgpu::binding {
         Converter conv(env);
 
         wgpu::Buffer b{};
-        uint32_t o = 0;
+        uint64_t o = 0;
 
         if (!conv(b, indirectBuffer) ||  //
             !conv(o, indirectOffset)) {
